+++ TripKit/Classes/server/TKAPIToCoreDataConverter.swift
@@ -237,12 +237,8 @@ extension Alert {
     
     hashCode = NSNumber(value: model.hashCode)
     title = model.title
-    if let startTime = model.startTime {
-      self.startTime = Date(timeIntervalSince1970: startTime)
-    }
-    if let endTime = model.endTime {
-      self.endTime = Date(timeIntervalSince1970: endTime)
-    }
+    startTime = model.fromDate
+    endTime = model.toDate
     if let location = model.location {
       self.location = SGKNamedCoordinate(from: location)
     }