+++ TripKit/Classes/server/TKBuzzInfoProvider.swift
@@ -77,6 +77,7 @@ extension TKBuzzInfoProvider {
                            completion: completion)
   }
   
+  // MARK: - Transit alerts
   
   /**
    Asynchronously fetches transit alerts for the provided region.
@@ -84,10 +85,14 @@ extension TKBuzzInfoProvider {
    - Note: Completion block is executed on the main thread.
    */
   public class func fetchTransitAlerts(forRegion region: SVKRegion, completion: @escaping ([API.Alert]) -> Void) {
-    
+    let paras: [String: Any] = [
+      "region": region.name,
+      "v": TKSettings.defaultDictionary()["v"] as! Int
+    ]
+
     SVKServer.shared.fetch(AlertsTransitResponse.self,
                            path: "alerts/transit.json",
-                           parameters: ["region": region.name],
+                           parameters: paras,
                            region: region)
     { response in
       let mappings = response?.alerts ?? []
@@ -99,18 +104,29 @@ extension TKBuzzInfoProvider {
    Asynchronously fetches transit alerts for the provided region using Rx.
    */
   public class func rx_fetchTransitAlerts(forRegion region: SVKRegion) -> Observable<[API.Alert]> {
+    return rx_fetchTransitAlertMappings(forRegion: region)
+      .map { $0.map {$0.alert} }
+  }
+  
+  public class func rx_fetchTransitAlertMappings(forRegion region: SVKRegion) -> Observable<[API.AlertMapping]> {
     let paras: [String: Any] = [
-      "region": region.name as Any
+      "region": region.name,
+      "v": TKSettings.defaultDictionary()["v"] as! Int
     ]
     
     return SVKServer.shared.rx
       .hit(.GET, path: "alerts/transit.json", parameters: paras, region: region)
-      .map { (_, _, data) -> [API.Alert] in
+      .map { (_, _, data) -> [API.AlertMapping] in
         let decoder = JSONDecoder()
-        guard let data = data, let response = try? decoder.decode(AlertsTransitResponse.self, from: data) else { return [] }
-        return response.alerts.map { $0.alert }
-    }
+        // This will need adjusting down the track (when using ISO8601)
+        decoder.dateDecodingStrategy = .secondsSince1970
+        guard let data = data else { return [] }
+        let response = try decoder.decode(AlertsTransitResponse.self, from: data)
+        return response.alerts
+      }
   }
+  
+  // MARK: - Accessibility
   
   /**
    Asynchronously fetches information about whether the provided region supports
@@ -172,6 +188,7 @@ extension SVKServer {
 
         do {
           let decoder = JSONDecoder()
+          // This will need adjusting down the track (when using ISO8601)
           let result = try decoder.decode(type, from: data)
           completion(result)
         } catch {