+++ TripKitUI/controller/TKAlertViewController.swift
@@ -15,22 +15,59 @@ import RxSwift
   import TripKit
 #endif
 
-@objc
-public protocol TKAlertModel {
+@objc public protocol TKAlert {
   var title: String? { get }
   var icon: SGKImage? { get }
   var iconURL: URL? { get }
   var text: String? { get }
   var infoURL: URL? { get }
+  var startTime: Date? { get }
   var lastUpdated: Date? { get }
 }
 
+// MARK: -
+
+/// This is a wrapper class that converts an instance of API.Alert into
+/// an instance that can conform to object protocol `TKAlert`.
+class TKAlertAPIAlertClassWrapper {
+  private let alert: API.Alert
+
+  init(alert: API.Alert) {
+    self.alert = alert
+  }
+}
+
+extension TKAlertAPIAlertClassWrapper: TKAlert {
+  var title: String? { return alert.title }
+  var iconURL: URL? { return alert.remoteIcon }
+  var text: String? { return alert.text }
+  var infoURL: URL? { return alert.url }
+  var lastUpdated: Date? { return alert.lastUpdated }
+  var startTime: Date? { return alert.fromDate }
+  
+  var icon: SGKImage? {
+    let fileName: String
+    switch alert.severity {
+    case .info, .warning:
+      fileName = "icon-alert-yellow-high-res"
+    case .alert:
+      fileName = "icon-alert-red-high-res"
+    }    
+    return TripKitUIBundle.imageNamed(fileName)
+  }
+}
+
+// MARK: -
+
 public class TKAlertViewController: UITableViewController {
   
-  private let disposeBag = DisposeBag()
+  @objc public weak var alertControllerDelegate: TKAlertViewControllerDelegate?
+  
   private weak var emptyAlertView: TKEmptyAlertView?
   
-  private var alerts: [TKAlertModel] = [] {
+  private let disposeBag = DisposeBag()
+  
+  public var alerts: [TKAlert] = [] {
     didSet {
       if alerts.isEmpty {
         insertEmptyAlertsView()
@@ -41,9 +78,6 @@ public class TKAlertViewController: UITableViewController {
     }
   }
   
-  public var transitAlerts: Observable<[TKAlertModel]>?
-  @objc public weak var alertControllerDelegate: TKAlertViewControllerDelegate?
-  
   // MARK: - View lifecycle
   
   override public func viewDidLoad() {
@@ -51,30 +85,24 @@ public class TKAlertViewController: UITableViewController {
     
     self.title = Loc.Alerts
     
-    if let navigator = navigationController,
-      let topCtr = navigator.viewControllers.first, topCtr == self {
-      let doneButton = UIBarButtonItem(barButtonSystemItem: .done, target: nil, action: nil)
-      doneButton.rx.tap
-        .subscribe(onNext: { [unowned self] in
-          self.dismiss(animated: true, completion: nil)
-        })
-        .disposed(by: disposeBag)
-      navigationItem.leftBarButtonItem = doneButton
-    }
-    
     tableView.rowHeight = UITableViewAutomaticDimension
     tableView.estimatedRowHeight = 150
-    tableView.register(TKAlertCell.nib, forCellReuseIdentifier: String(describing: TKAlertCell.self))
+    tableView.register(UINib(nibName: "TKAlertCell", bundle: Bundle(for: TKAlertCell.self)), forCellReuseIdentifier: "TKAlertCell")
     SGStyleManager.styleTableView(forTileList: tableView)
+  }
+  
+  public override func viewWillAppear(_ animated: Bool) {
+    super.viewWillAppear(animated)
     
-    transitAlerts?
-      .observeOn(MainScheduler.instance)
-      .subscribe(onNext: { [weak self] in
-        if let strongSelf = self {
-          strongSelf.alerts = $0
-        }
-      })
-      .disposed(by: disposeBag)
+    if let navigator = navigationController,
+      let firstCOntroller = navigator.viewControllers.first, firstCOntroller == self {
+      let doneButton = UIBarButtonItem(barButtonSystemItem: .done, target: self, action: #selector(doneButtonTapped(_:)))
+      navigationItem.leftBarButtonItem = doneButton
+    }
+  }
+  
+  @objc private func doneButtonTapped(_ sender: UIButton) {
+    dismiss(animated: true, completion: nil)
   }
   
   // MARK: - UITableViewDataSource
@@ -92,8 +120,7 @@ public class TKAlertViewController: UITableViewController {
       preconditionFailure("Index path refers to a non-existent alert")
     }
     
-    let alertCell = tableView.dequeueReusableCell(withIdentifier: String(describing: TKAlertCell.self), for: indexPath) as! TKAlertCell
-    
+    let alertCell = tableView.dequeueReusableCell(withIdentifier: "TKAlertCell", for: indexPath) as! TKAlertCell
     let alert = alerts[indexPath.row]
     
     // This configures the cell.
@@ -142,7 +169,7 @@ public class TKAlertViewController: UITableViewController {
 
 @objc public protocol TKAlertViewControllerDelegate {
   
-  @objc optional func alertViewController(_ controller: TKAlertViewController, didSelectAlert alert: TKAlertModel)
+  @objc optional func alertViewController(_ controller: TKAlertViewController, didSelectAlert alert: TKAlert)
   @objc optional func alertViewController(_ controller: TKAlertViewController, didTapOnURL url: URL)
   
 }